Implementation of __init__()
I added the __init__() method, linked the instance varibles with the predefiniton values, splited the price method so that I don't have to rewrite the same code in the init method. The init part isn't complete yet, because it's missing the if clause about the product not being in the predefinition. I alson change the pattern of the regular expressions used in the other methods.